Refactoring move
